+++ packages/components/src/date-field/index.ts
@@ -1,1 +1,4 @@
 export * from './DateField'
+export * from './DateSegment'
+export * from './DateInput'
+export * from './DateInputDivider'
