--- packages/components/src/date-input/index.ts
@@ -1,1 +1,0 @@
-export * from './DateInput'
