+++ packages/components/src/date-picker/DatePicker.tsx
@@ -8,11 +8,10 @@ import {
   type ValidationResult,
 } from 'react-aria-components'
 import { clsx } from 'clsx'
-import { DatePickerInputField } from './components/DatePickerInputField'
-import { DatePickerPopover } from './components/DatePickerPopover'
+import { DatePickerInputField } from './DatePickerInputField'
+import { DatePickerPopover } from './DatePickerPopover'
 import { Calendar } from '../calendar'
-import { DateInput } from '../date-input'
-import { DateSegment } from '../date-segment'
+import { DateInput, DateSegment } from '../date-field'
 import { FieldError } from '../field-error'
 import { Label } from '../label'
 import { Text } from '../text'
