+++ packages/components/src/date-picker/DatePickerPopover.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 import { Dialog, Popover } from 'react-aria-components'
-import styles from '../DatePicker.module.css'
+import styles from './DatePicker.module.css'
 
 interface DatePickerPopoverProps {
   children?: React.ReactNode
