--- packages/components/src/date-segment/index.ts
@@ -1,1 +1,0 @@
-export * from './DateSegment'
